requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.2
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
openai==0.27.8
pdfminer.six==20221105
portalocker==2.7.0
protobuf==4.23.2
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
python-dotenv==1.0.0
qdrant-client==1.2.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl==3.2.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
google-generativeai   